refactor: change RecordActivity to RecordFragment
