+++ application/app/src/main/AndroidManifest.xml
@@ -54,9 +54,6 @@
             android:exported="false" >
         </activity>
         <activity
-            android:name=".record.view.RecordActivity"
-            android:exported="false" />
-        <activity
             android:name=".main.MainActivity"
             android:exported="false">
         </activity>
