--- application/app/src/main/res/layout/activity_record.xml
@@ -1,37 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/main"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".record.view.RecordActivity"
-    android:orientation="vertical">
-
-    <Button
-        android:id="@+id/start_record_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:text="녹음 시작"/>
-
-    <Button
-        android:id="@+id/end_record_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_gravity="center"
-        android:text="녹음 중지"/>
-
-    <Button
-        android:id="@+id/move_stt_button"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="stt 테스팅 페이지로"/>
-
-    <Button
-        android:id="@+id/move_tts_button"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:text="tts 테스팅 페이지로"/>
-
-</LinearLayout>
