+++ application/app/src/main/res/menu/nav_menu.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android">
+    <item
+        android:id="@+id/recordFragment"
+        android:title="레코드"
+        />
+
+<!--    <item-->
+<!--        android:id="@+id/testFragment"-->
+<!--        android:title="테스트용"/>-->
+</menu>
