+++ application/app/src/main/res/navigation/nav_graph.xml
@@ -1,0 +1,16 @@
+<?xml version="1.0" encoding="utf-8"?>
+<navigation xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:id="@+id/nav_graph.xml"
+    app:startDestination="@id/recordFragment">
+
+    <fragment
+        android:id="@+id/recordFragment"
+        android:name="com.example.priend.record.view.RecordFragment"
+        android:label="fragment_record"
+        tools:layout="@layout/fragment_record"
+        />
+
+
+</navigation>
